requirements.txt
matplotlib
numpy
scipy
tqdm
pyyaml
easydict